refactor: modify module import
