+++ source/lambda/online/langchain_integration/chains/intention_chain.py
@@ -13,9 +13,9 @@ from langchain.schema.runnable import (
 )
 
 from common_logic.common_utils.constant import LLMTaskType,LLMModelType
-from ..llm_models import Model
+from ..chat_models import Model
 from .chat_chain import Iternlm2Chat7BChatChain
-from .llm_chain_base import LLMChain
+from . import LLMChain
 
 abs_dir = os.path.dirname(__file__)
