+++ source/lambda/online/langchain_integration/chains/retail_chains/retail_conversation_summary_chain.py
@@ -8,7 +8,7 @@ from langchain.schema.runnable import (
 
 
 from ...llm_models import Model
-from ..llm_chain_base import LLMChain
+from ..__llm_chain_base import LLMChain
 from common_logic.common_utils.constant import (
     MessageType,
     LLMTaskType,
